V-04 F-05 ready to deploy
