+++ src/hooks/useCheckboxFilter.js
@@ -7,11 +7,11 @@ const useCheckboxFilter = () => {
     const isFullTank = useSelector((state) => state.form.isFullTank);
     const isNeedChildChair = useSelector((state) => state.form.isNeedChildChair);
     const isRightWheel = useSelector((state) => state.form.isRightWheel);
+    const price = useSelector((state) => Number(state.form.price));
     const defaultCarValue = "Все модели";
     const defaultColor = "Любой";
     const [checked, setChecked] = useState(defaultCarValue);
     const [carColor, setColor] = useState(defaultColor);
-    const [tank, setTank] = useState(false);
 
     const check = (value) => {
         setChecked(value.name);
@@ -61,7 +61,100 @@ const useCheckboxFilter = () => {
         );
     };
 
-    return [checked, check, carColor, changeColor, changeRate];
+    const checkTank = () => {
+        // setTank(!value);
+        if (isFullTank.value === "false") {
+            dispatch(
+                formAction({
+                    isFullTank: {
+                        value: "true",
+                        name: isFullTank.name,
+                        id: isFullTank.id,
+                        price: isFullTank.price,
+                        form: "Да",
+                    },
+                    price: price + isFullTank.price,
+                }),
+            );
+        } else {
+            dispatch(
+                formAction({
+                    isFullTank: {
+                        value: "false",
+                        name: isFullTank.name,
+                        id: isFullTank.id,
+                        price: isFullTank.price,
+                        form: "Нет",
+                    },
+                    price: price - isFullTank.price,
+                }),
+            );
+        }
+    };
+
+    const checkChair = () => {
+        // setTank(!value);
+        if (isNeedChildChair.value === "false") {
+            dispatch(
+                formAction({
+                    isNeedChildChair: {
+                        value: "true",
+                        name: isNeedChildChair.name,
+                        id: isNeedChildChair.id,
+                        price: isNeedChildChair.price,
+                        form: "Да",
+                    },
+                    price: price + isNeedChildChair.price,
+                }),
+            );
+        } else {
+            dispatch(
+                formAction({
+                    isNeedChildChair: {
+                        value: "false",
+                        name: isNeedChildChair.name,
+                        id: isNeedChildChair.id,
+                        price: isNeedChildChair.price,
+                        form: "Нет",
+                    },
+                    price: price - isNeedChildChair.price,
+                }),
+            );
+        }
+    };
+
+    const checkWheel = () => {
+        // setTank(!value);
+        if (isRightWheel.value === "false") {
+            dispatch(
+                formAction({
+                    isRightWheel: {
+                        value: "true",
+                        name: isRightWheel.name,
+                        id: isRightWheel.id,
+                        price: isRightWheel.price,
+                        form: "Да",
+                    },
+                    price: price + isRightWheel.price,
+                }),
+            );
+        } else {
+            dispatch(
+                formAction({
+                    isRightWheel: {
+                        value: "false",
+                        name: isRightWheel.name,
+                        id: isRightWheel.id,
+                        price: isRightWheel.price,
+                        form: "Нет",
+                    },
+                    price: price - isRightWheel.price,
+                }),
+            );
+        }
+    };
+
+    return [checked, check, carColor, changeColor, changeRate, checkWheel, checkChair, checkTank];
 };
 
 export default useCheckboxFilter;
