+++ src/store/store.js
@@ -7,6 +7,10 @@ const store = configureStore({
         auth: authReducer,
         form: formReducer,
     },
+    middleware: (getDefaultMiddleware) =>
+        getDefaultMiddleware({
+            serializableCheck: false,
+        }),
 });
 
 export default store;
